디렉토리 정상화

+++ app/src/main/java/com/example/pda_project/models/Item.kt
@@ -7,7 +7,7 @@ import com.google.firebase.firestore.FirebaseFirestore
 data class Item(
     val productId: String,
     val itemName: String,
-    var amount: Long
+    var amount: Int
 ) {
     companion object {
         private val firestore: FirebaseFirestore by lazy { FirebaseFirestore.getInstance() }
